requirements.txt
requests
feedparser
PyMuPDF  # for PDF processing
pillow  # for image processing
google-generativeai  # for Gemini API
schedule  # for task scheduling
python-dotenv  # for environment variables